SolarPVUtil Jupyter notebook.
Instantiate a SolarPVUtil solution Jupyter notebook.
This takes the form of an ipynb notebook file, and a
solution/solarpvutil module which instantiates model/*
objects with parameters appropriate for the SolarPVUtil
solution.

This comes with a fairly trivial test case which checks
that the object can be instantiated. Previously we had a
proto-solarpvutil object in the tree but broke it without
noticing for some time. As we figure out how we're going
to surface results from the solutions, the test cases
should grow to cover it.

Also:
+ move advanced_controls and data_sources into model/*, having them
  at the top level made them look more important than they are.
+ Add PDS support in TAM module, somehow missed in the initial
  implementation.
+ Fix single source support in TAM forecast.
+ remove duplicate "Project Drawdown Model_Framework_and Guide.pdf"
  from top level directory.
+ update model/tests/pds_tam_per_region.csv and ref_tam_per_region.csv
  after fixing a problem discovered in the Excel model for the Asia
  region. Only 43 years of data were being used to do the curve fit
  not the full 46 years. This led to a bit over 1% difference in the
  2060 result, enough to make tests against the Python implementation
  (which always used 46 years of data) fail.