VMA refactor to use CSV files exclusively.
The initial Variable Meta-Analysis implementation took a
dataframe as an argument. This had a couple issues:
+ several of the VMAs had dozens of sources, leading to a giant wall
  of text buried in the Python code.
+ Researchers who are not familiar with Python need to be able to
  update the data sources. Putting it into a Python file means
  that some of them will be reluctant to even try.

Instead, make VMA objects read in a CSV file. CSV is a familiar
format for most climate researchers.

The unit tests ensure that reading from a file works, but then use
io.StringIO for most tests so the test data is kept with the test
code.