Support all six scenarios in SolarPVUtil.
SolarPVUtility_RRS_ELECGEN_v1.1d_27Aug18 contains six scenarios:
  + PDS3-16p2050-Optimum (Updated)
  + PDS2-15p2050-Drawdown (Updated)
  + PDS1-10p2050-Plausible (Updated)
  + PDS-16p2050- Optimum (Book Ed.1)
  + PDS-15p2050- Drawdown (Book Ed.1)
  + PDS-10p2050- Plausible (Book Ed.1)

Implement all six scenarios in solution/solarpvutil, and test
all of them in tests/test_excel_integration.py.

The Plausible scenarios exposed a bug in adoptiondata: the standard
deviation for ALL SOURCES and for a single source need to be calculated
over all available sources, but the standard deviation for a group
like Ambitious Cases should only include the sources in that group.
Add a unit test case to catch this, and fix it.