Remove advanced_controls soln_funit_adoption_2014
ac.soln_funit_adoption_2014 was added early on, but comes from
the same source as the first row in ht_pds_datapoints which
was added later.

This means it is also the first row of the pds_datapoints
passed in to unitadoption.py, the only user of
ac.soln_funit_adoption_2014

In accordance with the DRY (Don't Repeat Yourself) principle,
remove ac.soln_funit_adoption_2014 and use pds_datapoints[0]
in unitadoption.py. We're using the first datapoint, not 2014
specifically, as the project expects to eventually move the
time windows forward and not base all projections from 2014.