solution_xls_extract: infer a classname
We're expecting a lengthy period where the Excel models continue
to be developed in parallel with the Python, and a need to
repeatedly re-import the solutions and scenarios. Looking up
the classname we used last time is annoying, so infer a reasonable
classname from the filename.

The project has not been terribly consistent about filenames,
there will be multiple patterns for a number of the classes.

Also check that all of the fields parsed from the scenario are
actually consumed (i.e. outputted into the generated code), and
correct the source_before_2014/source_until_2014 error this
uncovered.