Add ConcentratedSolar solution.
Add a third solution: Concentrated Solar Power (aka CSP),
joining the earlier Utility-scale Photovoltaic and
Rooftop Solar solutions. The earlier two were constructed
by hand, but concentratedsolar was auto-generated using
tools/solution_xls_extract.

test_excel_integration.py tests a single scenario from
concentratedsolar. Attempting to test the others results in
a hard crash of Excel 2016 on MacOS. We've seen similar
crashes when an Excel file becomes corrupted, but the usual
fix for those cases is to open and save the file using a Windows
Excel version and that has not helped in this case. We'll keep
working on it.