metaclass_cache.py for caching class instances.
model/metaclass_cache.py is a Python Metaclass (i.e. like "type",
the superclass of object) which hashes arguments passed to __init__
and maintains a cache of allocated instances to their arguments.

Passing the same arguments to __init__ will return the same
object to the new caller. This is useful for:
+ classes where initialization is expensive...
+ ...which do not contain mutable state

In our current system this is of greatest applicability to tam.py.
Though lru_cache helps, tam.py initialization still shows up in
profiling as an expensive part of the system. The TAM also has the
property that many solutions overlap, using the same set of source
material. For example the energy solutions mostly use identical
TAM arguments, and we expect Land and Food to have similar levels
of overlap in the various solutions.

Benchmarking creation of all scenarios in currently finished
energy solutions:
    start = time.perf_counter()
    for scenario in concentratedsolar.scenarios.keys():
      _ = concentratedsolar.ConcentratedSolar(scenario=scenario)
    for scenario in landfillmethane.scenarios.keys():
      _ = landfillmethane.LandfillMethane(scenario=scenario)
    for scenario in solarpvroof.scenarios.keys():
      _ = solarpvroof.SolarPVRoof(scenario=scenario)
    for scenario in solarpvutil.scenarios.keys():
      _ = solarpvutil.SolarPVUtil(scenario=scenario)
    end = time.perf_counter()

Results in:
+ with lru_cache decorators removed:  92.5 seconds
+ with lru_cache (status quo):        56.6 seconds
+ with lru_cache and metaclass_cache: 36.3 seconds

Note that we previously did a largish refactor to move most arguments
out of methods and into __init__, anticipating creation of an object
cache like this. tam __init__, and to a lesser extent adoptiondata,
have been relatively expensive since the very early development of the
system, efforts to make them faster have been underway for some time.