Python round() != Excel ROUND in lifetime calc.
0.5 cannot be precisely represented as a floating point number,
it will be 0.499999. Python's default round() is fairly naive and
will round down to zero. Excel is more sophisticated and ensures
that 0.5 rounds up to 1.0 as humans expect.

This difference in handling was noticed while working on OnshoreWind,
where the replacement lifetime works out to 22.5 years and the
Python result did not match Excel. Use the decimal module to
implement handling similar to Excel, and add tests to demonstrate
the bug and verify the fix.