codegen: ht_*_datapoints[2050] is not constant.
tools/solution_xls_extract.py was treating ht_ref_datapoints
and ht_pds_datapoints as constants which had been entered by
the researcher, but only the 2014 row is such a constant.
The 2050 row is calculated based on the 2014 values and on
TAM Data, so scenarios which adjust the subset of sources
used to compute the TAM will also modify the datapoints
passed to helpertables.

Noticed while working on OffshoreWind, which is the first
solution which has scenarios where this makes enough
difference to cause tests/test_excel_integration.py to
fail.