unitadoption: remove in-place divide.
test_excel_integration is failing because
ua.soln_net_annual_funits_adopted() was being modified in
place, which modified the dataframe in the lru_cache.

As soln_net_annual_funits_adopted is used in a number of
places, add a test which calls those functions and checks
that the cached soln_net_annual_funits_adopted is not modified.
Verified that it catches the original problem as well.