tam: don't include 0.0 in the mean.
In Excel, the Mean computation is:
  SUM($C521:$Q521)/COUNTIF($C521:$Q521,">0")

The intent is to skip sources which are empty, but also means that
a source where the real data is 0.0 will not impact the Medium result.

See this document for more information:
https://docs.google.com/document/d/19sq88J_PXY-y_EnqbSJDl0v9CdJArOdFLatNNUFhjEA/edit#heading=h.yvwwsbvutw2j

We're matching the Excel behavior in the initial product.
This decision can be revisited later, when matching results
from Excel is no longer so strict a requirement.

Include a test case, showing that the mean of [0.0, 1.0] is
supposed to be 1.0. If this behavior is changed in the future,
the test case can be removed or updated to check for 0.5.