Add tools/topo
Script and source data to take a TopoJSON file of the world
(sans Antartica) and annotate it with Drawdown regions.

The intention is to use this TopoJSON file for interactive
maps in Jupyter.

tools/topo/world-countries-sans-antarctica.json is from
https://raw.githubusercontent.com/deldersveld/topojson/master/world-countries-sans-antarctica.json
https://github.com/deldersveld/topojson says, "All files are
believed to have been converted from sources that are licensed
for reuse or which are in the public domain."