tools/topo: reduce LAND mesh simplification.
toposimplify -p 1 resulted in a damaged mesh: an AEZ=10
region somewhere in Russia no longer fully closed, so the
fill algorithm in the Geo chart would determine that all of
the world's oceans were part of that AEZ=10 region. At the
time I thought this was normal, but it is now clear that
ocean is not supposed to be in any AEZ.

The flaw in this case is presumably in the original mesh not
actually being closed and relying on an intersection of two
segments to close that region. However as this is the only
GeoJSON AEZ source we've found, we're going to keep using it.

Make the TopoJSON simplification less aggressive, -p 0.75 instead
of -p 1. This results in a JSON file which is only about 4K larger.

Slightly updated Jupyter code to display a geo chart for a LAND
solution:

  source = alt.topo_feature('data/world_topo_with_aez.json', 'aez')
  alt.Chart(source).mark_geoshape(
      fill='#dddddd',
      stroke='black',
  ).transform_calculate(
      AEZ = 'datum.properties.AEZ',
  ).encode(
      tooltip=[alt.Tooltip('AEZ:O', title='AEZ')],
      color=alt.Color('id:O', scale=alt.Scale(
  	domain=list(range(0,19)),
  	range=['#eeeeee', '#B2EA78', '#99E27E', '#82DA84', '#6DD08A', '#59C78F', '#47BD93',
  	       '#37B296', '#2CA797', '#279C97', '#0077be', '#2D8691', '#347B8C', '#3A7085',
  	       '#40657D', '#445A73', '#464F69', '#46455D', '#453B52']))
  ).project('conicEquidistant').properties(
      width=600,
      height=450,
      title='AEZ'
  )