s_curve: initial implementation
Implements the first of three S-Curves, a logistic
sigmoid function. Includes test cases using data from
the Building Automation solution.

The second S-Curve is a Bass Diffusion model which is
expected in a future commit.

The third S-Curve is not believed to be used in any
solutions. We will only implement it if we find solutions
which use it.