interpolation: non-existent != fatal.
If a data source is specified and not found, it was being
treated as a fatal error in constructing the model. That isn't
always the case, for example when computing the PDS TAM if the
PDS source doesn't exist it is supposed to use the REF TAM for
that year.

Instead of raising ValueError return None, and adjust the test
case to match. The option of catching ValueError was considered
but didn't seem right, especially since this routine already returns
None for other similar cases.

Noticed while working on Smart Glass, which does this.