firstcost & operatingcost: adjust NaN handling.
Adjust handling of NaN, propogate it further through the models and
don't dropna() in firstcost. Based on experience with Alternative
Cement solution, which has a lot of N/A in its data.

This is likely to be refined further as we work through more
solutions and see the various ways there are holes in the data.