solution_xls_extract: default/custom ref ad & NaN.
There is at least one solution (Nuclear) which uses the Default
REF adoption in some scenarios and Custom REF adoption in other
scenarios. As there is only one solution so far, compared with
the large number of solutions with a mix of PDS adoption
mechanisms, we're only going to emit a decision tree in cases
where both mechanisms are present. If a solution only uses one
or the other, there will not be a conditional emitted.