operatingcost: support var/fixed conv_factor.
Excel supports separate conversion factors for fixed and
variable operating costs. We thought this was superfluous
and the two would always be the same, but there is at least
one solution (HeatPumps) where they differ.

As the function appears to be rarely needed, continue to support
passing in a single conversion_factor to be used for both values.
It isn't a good tradeoff to have 44 solutions pass the same factor
twice in order to support the one solution which needs them different.
If a tuple with multiple values is passed in, it will be used
as separate fixed and variable factors.

Also fix annual_breakout: it shouldn't depend on has_var_costs, which
is only true if all four of soln+conv var+fuel are present. Many of
the energy solutions have a fuel cost for the conventional technology
(coal, oil, etc) and no fuel cost for the renewable solution tech.