ac: shorten conv_2014_cost vma_title
From:
'CONVENTIONAL First Cost per Implementation Unit for replaced practices technologies'
to:
'CONVENTIONAL First Cost per Implementation Unit for replaced practices'
because the title in VMAinfo is:
'CONVENTIONAL First Cost per Implementation Unit for replaced practices/technologies'
and the difference in '/' and ' ' makes it not match.

First noticed with irrigationefficiency, but I don't understand why it
hasn't appeared before this. All solutions appear to have the slash in
the VMA name.