Redo solarpvutil using the public release.
solarpvutil was the first solution developed, and entirely by hand.
Data from solarpvutil is used in a large portion of the unit tests
for each module. The modules thus developed and the experience
gained while constructing the solution led to development of the
tools/solution_xls_extract.py code generator.

We're now replacing the manually created files with generated code.