Re-apply attribution comments after merge.
generated using: sed -i -e 's/$/  # bb/' <filenames>

The output of the code generator is likely approaching its final
form, as most of the existing Excel features work and new features
would be developed natively in Python. Even if solutions are
re-generated, I suspect most of the lines in the file would not
change.

I don't think it is reasonable for me to be shown as the originator
of the content of these files, I want to ensure the attribution is
correct. Add comments to the end of every line. A subsequent commit
can remove the comments using:

sed -i -e 's/  # bb$//' solution/afforestation/__init__.py \
  solution/forestprotection/__init__.py solution/peatlands/__init__.py \
  solution/silvopasture/__init__.py \
  solution/temperateforests/__init__.py \
  solution/tropicalforests/__init__.py \
  solution/tropicaltreestaples/__init__.py

There are more files than this, I am sure. This is the first set.