UI: add links to Discourse, doc, video.
Add a link to the Discourse community site. The URL for the site is
not final, but it will do for now.

Add a link to the Getting Started doc (a bit.ly link so we can move
it freely later).

Add a link to a YouTube video for how to get started.