solarhotwater: manually edit VMA summary
The ScenarioRecord values in Excel are stored with 10 digits of
precision, while the VMA summary values are stored with 11.
SolarHotWater's lifetime was deliberately chosen by the research
fellow to be 20 years, but the difference in the number of digits
makes it 19.999999999999 years and the FirstCost tab calculations
round down. Changing from stored values in the advanced controls
scenario to VMA mean changed the behavior of FirstCost enough
to break test_excel_integration. Manually adjust the VMA values
to be the same as the advanced controls values were before.