ac: suppress value_from_excel message
Warning messages appearing in the UI are not a good way to flag an
error in a solution file. Implement an exception instead, suitable
for hooking up in a unit test laster. We'll have to fix all of the
solutions which have this problem (currently believed to be Cars
and Insulation) before adding that unit test.