UI: Compact interface, one solution at a time.
Restructure the UI to include the Overview list of solutions and
treemap+donut chart graphics as a tab alongside the other tabs.

Replace the checkboxes with radio buttons, allowing only one solution
to be selected at a time. This is kindof sad, as the ability to
show results from multiple solutions is sometimes useful, but interferes
with the goal of being an editing environment for researchers.