modelmap: suppress ns0 namespace.
Jupyter Notebook's display(SVG()) method does not tolerate having
an ns0: tag prepended to SVG nodes. It wants a default namespace.
The original SVG (generated by Omnigraffle on a Mac) does not have
prefixes and uses a default namespace, but elementtree adds them
when it generates the output.

I don't see a way to make elementtree use a default namespace.
We could change it from ns0 to something else using register_namespace,
but not omit the prefixes altogether.

So we postprocess the generated XML output to remove ns0 prefixes and
add an "xmlns=..."

The resulting SVG works in Notebook, and still works in Jupyterlab.