UI: support Jupyter Notebook for Vega.
Try importing vega.vega_widget, the nbextension to support Vega in
Jupyter Notebook, and use it as a heuristic of whether we're running
within JupyterLab or not.

This isn't a great heuristic, as someone could easily be running
Notebook without having vega_widget installed, but it is an initial
step.

Jupyter Notebook support is being worked on mainly to be able to use
Voila (https://github.com/QuantStack/voila)