UI: provide a separate Drawdown.ipynb for Voila.
To avoid scrolling the viewport when running within the
Notebook we need a %%javascript cell, but JupyterLab can
not run bare JavaScript due to its security model.

Give up on having one ipynb file with heuristics to guess
whether we're running in the Notebook or Lab. Provide a
separate voila/Drawdown.ipynb file, and explicitly tell
the ui.JupyterUI object which environment it should use.