tests: test_excel_integration, v3.
test_excel_integration runs the original Excel implementation of the
solutions, the Python implementation, and compares the results
of each step of the model to show that Python gets the same result
for the same reason as the original.

This is v3, the second substantial rewrite of the original test.

****** v1 ******
v1 of this test was created very early in the project, in the second
week. It used xlwings to perform all operations: at that time, we were
exercising the Python code using a local Flask webserver, with VBA code
in a modified Excel spreadsheet to fetch results via HTTP.

v1 worked well for the first several solutions, but xlwings to remotely
operate Excel is an enormously complex way to run a test. It would
sometimes just stop for an extended period of time until the operation
timed out, and the more operations the more likely at least one of
them would wedge and time out. By the time we got to ~10 solutions,
the test would get a flake failure on almost every run. The failure
would pass on re-run.

****** v2 ******
At the end of February 2019 in commit 1223913e6be5fd8a815d10e145288f06a37a20ba
test_excel_integration was rewritten to be less reliant on xlwings.
xlwings would be used to make Excel run the VBA code to load a scenario,
then save the resulting file. xlrd would then be used to read the
test values from this file and do the test asserts. The much reduced
use of xlwings made the test sufficiently reliable that flakiness has
not been noted even with an increased number of solutions.

There are two main issues with v2:
1. As the number of solutions has increased from ~10 to the current 68,
   the run time of the test case has grown to about 4 hours on a Macbook
   Pro. Realistically this means leaving it to run overnight, especially
   because it repeatedly starts Excel which causes brief interruptions
   in whatever one is trying to work on in the foreground.
2. The requirement to run Excel as part of the test means it cannot run
   as part of the Continuous Integration suite on the Linux build
   systems. This could probably have been addressed running Windows or
   MacOS in a VM, but life is too short to spend it on that kind of
   nonsense.

Not being able to run the test in CI means we've broken it several times
by accident and only realized it later.

****** v3 ******
v3, the version being committed here, decouples xlwings for preparation
from the actual running of the test. The code to run Excel, use xlwings
to run the VBA code to load each scenario, and save the results has been
moved to tools/integration_test_xls_extract.py. It is run as:
 python tools/integration_test_xls_extract.py --solutiondir=solution/solarpvutil

integration_test_xls_extract.py saves the results to
solution/<name>/testdata/expected.zip, a zip file of CSV files:

$ unzip -l ./solution/solarpvutil/testdata/expected.zip
Archive:  ./solution/solarpvutil/testdata/expected.zip
  Length      Date    Time    Name
---------  ---------- -----   ----
      491  08-08-2019 16:55   PDS-10p2050-Plausible (Book Ed. 1)/License
     1741  08-08-2019 16:55   PDS-10p2050-Plausible (Book Ed. 1)/Basic Controls
    68411  08-08-2019 16:55   PDS-10p2050-Plausible (Book Ed. 1)/Advanced Controls
    11041  08-08-2019 16:55   PDS-10p2050-Plausible (Book Ed. 1)/Detailed Results
    ...

tools/integration_test_xls_extract.py has to be run on a Mac or Windows
system with Excel installed, but the resulting expected.zip files are
checked in for any system to use.

test_excel_integration.py gets its expected values from these zip files,
and does not run Excel at all. The resulting test is 8x faster,
completing in half an hour what v2 took four hours to run.

The v3 test can also run on Linux, as it does not use Excel any more.

This CL is being checked in with ConcentratedSolar and
TropicalTreeStaples marked as skipped. It turns out we had checked in
new Excel files without updating test_excel_integration, which caused it
to quietly skip those tests due to the missing Excel file. We broke those
two solutions without realizing it. A future commit is planned to fix the
tests, and we made sure that v3 will not quietly skip a test if expected.zip
is missing. It will just fail instead.