Mark test_excel_integration as being slow.
v3 of test_excel_integration is 8x faster than v2, but
still takes half an hour which is 3x longer than the
rest of the unit tests.

Previously we'd marked the test as "integration" but this
is misleading as we have other integration tests which
were not being treated specially (nor even marked).

Mark test_excel_integration as "slow", and don't run it
from "make test"

Also take the next step of declaring the valid set of test
marks, and fail if an unknown mark is used, to catch typos.