FaIR: move into co2calcs
We initially wrote FaIR as a seperate model object because we
expected to set useMultigas=True and pass in CO2, CH4, N2O, HFC,
etc. We're not going to be able to do that right away as we don't
have all of the data which would be required, so we're currently
using useMultigas=False and passing in CO2eq.

Having FaIR be in a separate module is now making it harder to
develop, without benefit, so move it into co2calcs as everything
it does is based on CO2-eq.

If the day comes when FaIR takes inputs for CH4/N2O/etc, it can
be moved back out into its own module.