SVG: work around Notebook issues.
Work around https://github.com/ipython/ipython/issues/1866 in
Jupyter Notebook by prepending a unique slug to the id fields within
the SVG.

This also requires updating any filter, marker-start, and marker-end nodes
which have a 'url(#id)' reference to refer to the beslugged version.

If ever you see an issue where the SVG on the Model tab is fine but the
rest are broken in some way, suspect a cross-SVG ID referencing issue.
The Model tab is the first one rendered and will "win" any such
conflict.

JupyterLab renders SVGs within an iframe and is not impacted. Only
Notebook (and Voila) really need to have the IDs uniquified. As it
doesn't hurt JupyterLab to do this, we do so unconditionally.