aez: less dependence on global regions
Make aez use definitions in model/dd.py so as to be less sensitive
to knowing that it is using the global regions, to make regional
models easier.

There is still one remaining: it knows about a 'Global' region, which
is not region[0]. That will need more work.

Also reformat for style guide.