tropicaltreestaples: solution_aez_matrix.csv
solution_aez_matrix.csv for which AEZs were appropriate
was not correct for Tropical Tree Staples. Add a test,
and correct it. The test may be fragile as it checks
the specific values in data/land/, but right now there
isn't much alternative.

Tropical Tree Staples still doesn't pass test_excel_integration.py
as it has a dynamic Custom PDS Adoption which isn't implemented,
but it is one step closer now.