*forests: use customadoption datapoints.
tropicalforests and temperateforests use a CustomTLA for the
book edition scenarios, and the default TLA for newer scenarios.
The auto-generated version of the solutions didn't cope with this,
we skipped the non-CustomTLA scenarios in test_excel_integration.py.

Remove CSV files extracted from Excel, instead convert to
the customadoption datapoints feature for linear prediction.
This allows all scenarios to pass in test_excel_integration.py.