tam: confine regional sources to that region.
We support having unique data sources per region using a nested
'Region: NAME' structure, like so:
    data_sources = {
        'Baseline Cases': {
            'A': filenameA,
            'B': filenameB,
        },
        'Region: OECD90': {
            'Baseline Cases': {
                'C': filenameC,
            },
        },
    }

However if filenameC contained data for the World region, we would
add it to World. The regional data_sources provided a way to select
unique sources, not strictly keep them separate.

This was harmless until the work of the 2019 cohort of research fellows,
when the energy models started using a very different set of sources
for World results as for regional.

Implement more strict separation of data, and add a test to ensure it
works.