VMA: share data, move def into __init__.py
The set of VMAs for a solution is a part of the implementation of that
solution, move the definition out of a CSV file and into __init__.py.
The data for the VMA, including Fixed Mean/High/Low for non-open
sources, is best handled in CSV.

If the data for a VMA already exists in data/*, refer to that common
source instead of making a new CSV file within the solution.

Convert solarpvutil to the new implementation & verify that currently,
old and new ways of defining VMAs both work. We'll remove the old way
(by deleting model/vma.py:generate_vma_dict()) later after all solutions
have been updated.