VMA: remove postprocess, revamp has_data.
Remove the postprocess argument. It was intended to allow things like
having solarpvutil and solarpvroof share a VMA definition (as the two
together add up to 100%), but we're not using it in practice.

Revamp has_data, and remove the explicit argument for it. If there
is no data supplied when the VMA is created, then treat it as not having
data and return np.nan for the mean/high/low.