excel_fair_results: generate animations.
Generate animation of the impact on temperature for each sector.
This initial commit calculates the impact of each sector individually.
This isn't exactly correct as FaIR results are non-linear, the
impact of the sum of all sector emissions reductions will be greater
than the temperature impact of each sector individually.

A future commit may calculate the total temperature impact and then
apportion it by the faction of emissions reduction from each sector.