firstcost: remove regional costing and tests.
firstcost added support for computing cost per region some time
ago. This is a reasonable idea, but it is not used by any of the
solution models and is incomplete as only the first cost supports
it.

This should be done in the future when we're ready to use it, and
done for keepsies across all solutions not by adding conditionals
throughout the implementation.