FaIR animation: use fraction of total temperature.
FaIR is non-linear. The peak temperature and time spent at elevated
temperature are both factors in the model, accounting for feedback
effects.

Therefore, calculating the temperature impact of the emissions reduction
from each sector one at a time has less impact on temperature than the
that of all sectors together would. The whole is greater than the sum
of the parts.

+ calculate the overall temperature impact of all sectors together
+ apportion a fraction of the temperature reduction to each sector,
  based on the fraction of emissions reduction it achieves.