codecov.io: set patch coverage to 80%
By default, codecode.io requires the code coverage of the lines actually
changed by a commit to be greater than or equal to the coverage in the
overall project, which is currently >93% for this project.

Setting the bar that high has a couple issues:
1. It tends to dissuade people from contributing, if they cannot get the
   CI tests to pass.
2. It tends to encourage people to merge failing commits, which is not a
   good habit to be in.

We set a fixed target instead of the coverage of the overall repository.
Start with 80% and see how well that works.