Run workflow on pull request.
Codecov has to be skipped because the secret isn't available in pull
requests from forks, but we want to run the rest of the workflow.
This means that coverage will not be updated until the next pull_request
from the main repository... which is not ideal, as we may not realize it
was the earlier pull_request which impacted coverage.

codecov.io understands that this is an issue which needs a better
resolution: https://github.com/codecov/codecov-action/issues/44