codecov.io: coverage threshold 0% -> 0.1%
By default, codecode.io requires the code coverage of a pull request to
not reduce the coverage of the project, at all. This is proving
problematic, especially for changes which delete code, because the
percentage coverage of the remaining code will be slightly changed.

Setting the bar that high has a couple issues:
1. It tends to dissuade people from contributing, if they cannot get the
   CI tests to pass.
2. It tends to encourage people to merge failing commits, which is not a
   good habit to be in.

Allow the coverage to be ever so slightly reduced without failing the
commit. Start with 0.1%, as we're seeing changes failing due to 0.01% to
0.06%.