Close files opened by mkstemp
The first argument returned from tempfile.mkstemp is the os.open fileno,
which has to be explicitly closed or it will leak. When trying to import
a solution with a large number of scenarios, like the Drawdown2020
Landfille Methane solution, it can fail with too many open
file descriptors.

Close all files when finished with them.