ad: handle Medium for a single source.
Some time ago we implemented handling for the Medium computation,
to get the same answer as Excel's implementation:
SUM($C46:$Q46)/COUNTIF($C46:$Q46,">0")

The intent is to skip sources which are empty, but also means that
a source where the real data is 0.0 will not impact the Medium result.

The Excel behavior is somewhat more subtle, though: if the data is only
a single source it is used directly without computing a Mean.
The Nuclear solution in Drawdown 2020 ran into this issue. Implement
support and unit test.