Switch survey-health workflow to create a PR.
The master branch is protected in this repository, and the GITHUB_TOKEN
does not have the rights to push to it. Though we could create a
Personal Access Token for someone who does have that permission, the
failure mode if the bot does something bad is harder to recover from if
people cloned master in the time it was broken. We'd like to preserve
the chance for review before merging.

Switch from stefanzweifel/git-auto-commit-action@v4.1.1 to
peter-evans/create-pull-request@v2.4.4 to create a pull request
instead.