survey health: push to a different branch
We tried having the Workflow create a Pull Request, but (by
github's design) an Action doesn't trigger other Actions so
the PR does not run the continuous integration tests.

Instead, push to a specially named branch and expect one of the
maintainers to manually create a Pull Request. We'll see if this
works well enough.