Extend Selenium test, run in CI.
"smoke" is fairly well known as a first test to run (to check if it
smokes), but this one is focussed on just the Web UI not testing
the overall project. Rename it to web_ui.

The start.sh and stop.sh were a good way to get started, but made the
assumption that the voila for the test would be the only one running.
When run on developer workstations this isn't always true. Move voila
start into Python, capture the URL of the instance it ran, and shut it
down after the test.

Extend the test case to check for UI elements, to click the checkbox
to render one solution, and to check that the solution rendered.