ca: inferred world and out-of-scope datapoints.
Some solutions (like Afforestation) use a set of datapoints for
the major regions 'OECD90', 'Eastern Europe', 'Asia (Sans Japan)',
'Middle East and Africa', and 'Latin America', and have modified
the spreadsheet to make the World be a sum of these regions.

Implement this by using NaN as a magic value int he datapoints,
if the primary region's datapoints are NaN that means to sum
dd.MAIN_REGIONS to supply it.

Additionally, Afforestation supplies datapoints back in the 1990s
because the source data supplied data back to the 1990s, but the
model should only produce data for 2012-2060.