ca: adoption limits apply before averaging.
When there are multiple custom adoption data sources, there is
an adoption limit, and the average is called for, the limit
is supposed to be applied to each individual source before the
average is computed.

Otherwise it is possible to take an average where one source is
substantially above the limit but the average is not. That average
will be too high because it included out of range data.