ca: least_sq and maximum value.
1. When there are more than two datapoints:
  'datapoints': pd.DataFrame([
     [1990, np.nan, 96.8914, 44.1631, 106.1562, 16.5338, 14.9281],
     [2000, np.nan, 97.2555, 44.2605, 107.5489, 16.7251, 15.2771],
     [2005, np.nan, 97.6355, 44.3794, 109.8028, 16.9607, 15.6649],
     [2010, np.nan, 97.8700, 44.4540, 111.5210, 17.1760, 16.0730],
     [2014, np.nan, 98.1783330003811, 44.5558196042818, 113.7890763825080,
         17.4259450169749, 16.5131623025461],
     ], columns=["Year", "World", "OECD90", "Eastern Europe", "Asia (Sans Japan)",
         "Middle East and Africa", "Latin America"]).set_index("Year")

Solutions encountered thus far implement a piecewise interpolation:
1. Linear interpolation between 1990 and 2000
2. Separate linear interpolation between 2000 and 2005
3. Separate linear interpolation between 2005 and 2010
etc

However, the Excel model also implements a least squares fit through all
of the datapoints, as implemented by the Excel TREND() function.
Afforestation in the Drawdown 2020 solutions is the first place this
was noted to be used. Implement the same function using numpy polyfit.

2. maximum
Afforestaton in Drawdown 2020 also imposes a maximum value for one of
its sources, implemented by modifying the formulae of the cells in that
source. Implement a similar function in the Python model.