ca: bug-for-bug compat, source w/ no limit
In Excel, Custom PDS sources must each individually impose the
total_adoption_limit within the formula of the cell. This usually
takes the form of a MIN() statement:
  MIN($A$1, <rest of formula>
where $A$1 is the cell containing the total addressible market or
total land area for this solution.

We added support for applying the total adoption limit prior to
computing the Average/High/Low in
20a44a307b7044b412fd841a3558e9c637760fe4 because Afforestation
in Drawdown 2020 requires this behavior.

A number of solutions neglected to do this in one or more of their
custom sources, but there are two where the lack of applying
the limit impacts the results and now causes test_excel_integration
to fail: Perennial Bio Energy and Solar Hot Water.