ac: additional VMA values
Provide a way to place additional VMA values by title in the advanced
controls JSON file, necessary for solutions like Perennial Bioenergy
which use Additional Variables in their scenario record.

This mechanism is preliminary and will likely change in the future,
because it results in two different mechanisms for VMA values:
1. well-known VMAs like sequestration rate are stored by varname:
    "seq_rate_global": {
        "value": 1.183125,
        "statistic": "mean"
    },

2. Additional variables are stored by title:
    "vma_values": {
        "Future adoption (million hectares)": 0.591370993231854
    },

It seems like we should be able to get something much more consistent.