survey-health: switch to pull-request
We'd tried using peter-evans/create-pull-request previously,
using the default github actions token, but it did not run
the other pull request workflows. We switched to instead try
to push to a branch and manually create the PR, but that isn't
really working better.

Switch back to peter-evans/create-pull-request, and give it
a Personal Access Token scoped to 'repo' from the DentonGentry
account. This should let it run the other pull_request workflows.