Update Bamboo and Afforestation to 3/2020.
Commit 53c13263cced07e961fb2fa7382676b190cd2494 causes Bamboo
and Afforestation to fail the integration test. Added corrective
steps to the "Changes in Model spreadsheets to import for
test_excel_integration" document at
https://docs.google.com/document/d/1OiKg3_OOGjYOUdnHTQuZggsko5n31qv_YV4h77E3LHk/comment

Updated both solutions to the 3/2020 release, with 8 Thermal
Moisture Regimes.

Note that we're starting a new practice of storing the ZIP files
in LFS. Eventually these are going to be deleted, once we can
discontinue Excel altogether, LFS means they won't need to be in
the git object store and won't take local disk space for devs.