tla: support a fixed_value from advaced_controls.
Though the Custom TLA data in Excel is a NxM block of cells which
in theory could vary the land area per year, perhaps trying to model
impacts of warming on land use, in practice all solutions to date
use a single fixed value stored in the ScenarioRecord.

Support this use case in CustomTLA and the code generator.