fc: 0.0 is not NaN.
In Excel, NaN^0 == NaN. In Python, NaN^0 == 1.
We match the Excel behavior. See First Cost Handling of NaN data in:
https://docs.google.com/document/d/19sq88J_PXY-y_EnqbSJDl0v9CdJArOdFLatNNUFhjEA/edit#

Previously, we has implemented that both NaN and 0 would map to zero,
based on the behavior of the Nuclear solution. However, the Mangrove
Protection solution definitely does not expect 0 to be NaN, and the
updated Drawdown 2020 Nuclear solution doesn't either. I suspect this
was never correct, but worked around a problem which we have
subsequently fixed. Remove the behavior and the test for it.