oc: Use land units for annual_breakout.
RRS solutions have a notion of implementation units separate from
functional units.

Land solutions only have a single Land unit. We treat land units as an
funit, which is mostly correct, but the costs in the annual_breakout
were using iunits (which is always zero for land solutions). The
3/2020 version of Regenerative Agriculture is the first solution where
this was noticed.