ac: Support customadoption high/low multipliers.
Excel allows each ScenarioRecord to have a different multiplier
of the number of standard deviations above/below the average to
use for CustomAdoption. We had implemented this support in
model/customadoption but were passing it in from the solution
__init__.py file, as we had never found a solution which varied
them between scenarios.

Forest Protection in the Drawdown 2020 update does, so implement
the functionality.