ht: single_source overrides all valid years
In the case of Existing Adoption Prognostications with a single source,
we were only overriding the output starting in the year of the first
datapoint. This wasn't correct, and there isn't a reason to do so:
solutions now routinely set 2018 as their first datapoint, but the
years 2014:2017 still have data in the single_source.

No existing solutions in test_excel_integration were impacted by this
change, nothing depended on the old behavior.

Add a unit test for this specifically.