test_excel_integration: Excel_one_cent abs=0.01
The "Excel_one_cent" mask says to mask off any element in a DataFrame
which is less than 0.01 (i.e. one penny), as several of the solutions
involve multiplying very large $/unit by very small numbers of units.
The last couple bits of precision, amounting to tiny fractions of a
penny per year, is not relevant.

Here we add an absolute difference of 0.01 to Excel_one_cent comparisons.
If the difference is less than one cent, call it equal. By default
pytest uses an absolute threshold of 1e-12 and relative threshold of
1e-6. If an abs= argument is provided (as we do here 1e-2), and no
rel= argument, pytest will set rel=None and only check the absolute
difference. That isn't what we want, we do still want the relative
difference check so we set it to 1e-6 explicitly.