Update Electric Vehicles for Drawdown 2020.
EVs-Modelv1.1c-Apr2020.xlsm
https://drawdown.org/solutions/electric-cars

We've long contemplated retaining trimmed down versions of the
Excel spreadsheets, particularly for any new sheets added by the
researcher to calculate something unique for the given solution.
A spreadsheet really is a good tool fo that job. Electric Vehicles
is the first solution where we do this for keepsies: EV Sales
data is left in the Excel file and read in by the Python code.