ua: use ref_base_adoption for bug_cfunits_double_count
bug_cfunits_double_count was initially implemented using
the first row of soln_pds_funits_adopted, however this only
works if the first row is 2014. For the 3/2020 versions of
the solutions in the Drawdown Review the base year is now
2018, so this no longer works.

However we added an explicit ref_base_adoption in AdvancedControls
for these solutions. If ref_base_adoption is present, use it to
implement bug_cfunits_double_count. If not, fall back to the
previous method using the first row of soln_pds_funits_adopted.