Update Electric Trains for Drawdown 2020.
TrainFE-RRSv1.1b_23Oct19.xlsm
https://drawdown.org/solutions/electric-trains

Custom PDS Adoption data had been implemented on some added sheets
n the Excel workbook. As we have done for a few preivous solutions
like electricvehicles and hybridcars, we have chosen to check in an
XLSX file containing these sheets and read this file in from Python.
Excel really is a good tool for unique one-off computations as part
of a model.

The Book Ed. 1 scenarios have been renamed to *.broken because the
Custom PDS Adoption scenarios they use have been renamed in the
current model. In Excel, if a scenario specifies a name which does
not exist it will fall back to taking the average of all scenarios.
In Python we have chosen not to implement this because a lot of the
time the mismatch in scenarios is due to a typo and the human
researcher had no idea that Excel was taking the average instead.

So in Excel the Book Ed. 1 scenarios do get a result, but not what
the researcher intended, and in Python we've chosen not to implement
this behavior.