excel_integration: split expected and actual masks
We allow passing in a mask of values not to be compared, supporting
two distinctly different cases:
+ mask off cases where we know the result will differ, for example
  in cases where there is catastrophic cancellation in one or more
  years of a result. Though Excel may come up with 10e-11 and Python
  with 10e-12, we don't really care if the mean value is 10e6.
+ mask off cases where we know something about Excel's handling differs,
  for example in Excel NaN to the zeroth power is NaN where in Python
  it is 1. We handle these by passing in a string name of the desired
  mask, because it is computed based on values read in from
  expected.zip.

Unfotunately we now have cases where one solution needs to pass in
a DataFrame of booleans while another needs to pass in a string name.
Separate these two cases into their own arguments, allowing both
to be passed in.

Also, an earlier whitespace reformatting is harshing my mellow.
Unwrap the lines, just let them be long.