Remove obsolete energy data files.
Energy solutions all share the same TAM and a number of VMAs.
We previously consolidated the data files into a common data/energy
when importing the 2019 version of the solutions. We recently
updated the solutions to the 3/2020 versions from Excel, which
changed these data files. Before trying to consolidate the updated
versions, remove the ones no longer used.