GHG Accounting
(1) Updated the greenhouse gas accounting method for CO2 and CH4. This way we can show results in both individual gas units and CO2-equivalent units. (2) Introduced a module for N2O. (3) Updated the Global Warming Potential (GWP) conversions in the emission factors. (4) Included a tutorial for how to use the new features with example solutions Improved Rice and Landfill Methane.

- Kristina Colbert