address new edge case in adoption data
There were actually two minor bugs that cancelled each other out, until now:
(1) The Excel code that computes standard deviation treats groups with one column differently than single columns.  Added a fix to that.
(2) In the Electricity Generation models, the code for standard deviation did not treat the maximum group as a group, so if it was targeted, it was treated as a single column.  Added a quirks option to account for this case.
With Recycled Metals, we for the first time have a Maximum group with two columns (though one of the columns is empty), which exposed the first error; fixing that exposed the second one.