Delete extra scenarios; add PDS1, PDS2, PDS3
This is one of those changes where you wonder if it was worth the work, but...

To do integrations I will need to conveniently be able to get the right scenarios to work with for each solution.  That could be maintained in a list outside the system, but I decided to go through, and for each solution, assign values PDS1, PDS2 and PDS3 to name the most recent scenario of each type.  This will have to be maintained manually as new scenarios are published, but I figure this it be helpful in other situations as well (like a newcomer trying to get the code working for the first time).

Since I was looking in depth at the scenarios of each and every solution, I decided that would be a good opportunity to remove the "extras" while I was at it.