Allow TAM to be overridden by scenario
To support more dynamic scenario exploration, we are moving to put more options in the scenario that are currently hardwired in the solution init code.
I also used this exercise to centralize some code at the Scenario class level, rather than having it repeated in each init file.  This should become pattern across much of the init code in the future.