Add bound_correction to VMA
There is a parameter to VMAs that we weren't extracting or using: it sets the low output to min instead of mean-3sd.  It is intended to avoid negative low values where they don't make sense.  I have added this parameter, as well as made all "behavior" parameters (low_sd etc.) available for override on the main avg_high_low() calculation.

BTW, I suspect the absence of this parameter may be the source of some model test failures.